server/requirements.txt
# Server dependencies for Docker
# These are the minimal dependencies needed to run the server

pydantic>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
openenv-core>=0.2.2
requests>=2.31.0
python-dotenv>=1.0.0
